Refactor: NoteDetailViewModel and NoteDetailScreen and deleteNoteById input type to long

+++ app/src/main/java/ir/fatemelyasi/note/model/noteLocalDataSource/NoteLocalDataSource.kt
@@ -13,7 +13,7 @@ interface NoteLocalDataSource {
     suspend fun insertNote(note: NoteEntity) : Long
     suspend fun updateNote(note: NoteEntity)
     suspend fun deleteNote(note: NoteEntity)
-    suspend fun deleteNoteById(noteId: Int)
+    suspend fun deleteNoteById(noteId: Long)
     suspend fun deleteAll()
     fun searchNotes(query: String): Flow<List<NoteEntity>>
     fun getFavoriteNotes(): Flow<List<NoteEntity>>
